moved login user selector to it's own component
